add check-if-exists

+++ scripts/devOpsScripts/delete-resources.sh
@@ -53,13 +53,21 @@ echo "Found $appServiceDiagnosticName diagnostics in $resourceGroupName in $loca
 postgresSqlDiagnosticName=$(az monitor diagnostic-settings list --resource-group "$resourceGroupName" --resource "$postgresServerName" --resource-type Microsoft.DBforPostgreSQL/flexibleServers --query "[0].name" -o tsv)
 echo "Found $postgresSqlDiagnosticName diagnostics in $resourceGroupName in $location"
 
-echo "Deleting $appServiceDiagnosticName diagnostics in $resourceGroupName in $location"
-# delete the app service diagnostic so that it does not persist through resource group deletion and conflict with TF plan
-az monitor diagnostic-settings delete --resource-group "$resourceGroupName" --resource "$appServiceName" --name "$appServiceDiagnosticName" --resource-type Microsoft.Web/Sites
+if [[ ${#appServiceDiagnosticName} -eq 0 ]]; then
+  echo "Deleting $appServiceDiagnosticName diagnostics in $resourceGroupName in $location"
+  # delete the app service diagnostic so that it does not persist through resource group deletion and conflict with TF plan
+  az monitor diagnostic-settings delete --resource-group "$resourceGroupName" --resource "$appServiceName" --name "$appServiceDiagnosticName" --resource-type Microsoft.Web/Sites
+else
+  echo "Skipping delete for appServiceDiagnosticName diagnostics"
+fi
 
-echo "Deleting $postgresSqlDiagnosticName diagnostics in $resourceGroupName in $location"
-# delete the PostgresSQL server diagnostic so that it does not persist through resource group deletion and conflict with TF plan
-az monitor diagnostic-settings delete --resource-group "$resourceGroupName" --resource "$postgresServerName" --name "$postgresSqlDiagnosticName" --resource-type Microsoft.DBforPostgreSQL/flexibleServers
+if [[ ${#postgresSqlDiagnosticName} -eq 0 ]]; then
+  echo "Deleting $postgresSqlDiagnosticName diagnostics in $resourceGroupName in $location"
+  # delete the PostgresSQL server diagnostic so that it does not persist through resource group deletion and conflict with TF plan
+  az monitor diagnostic-settings delete --resource-group "$resourceGroupName" --resource "$postgresServerName" --name "$postgresSqlDiagnosticName" --resource-type Microsoft.DBforPostgreSQL/flexibleServers
+else
+  echo "Skipping delete for postgresSqlDiagnosticName diagnostics"
+fi
 
 echo "Deleting resources in $resourceGroupName in $location"
 # Delete the resource group
